Remove legacy CI build step
I'm not sure if we wan't to maintain a "legacy" CI step anymore. If we wish to test against backwards compatibility we should probably have a build step for each minor version we care about. This will more clearly allow us to see which versions are broken with each change.